requirements.txt
langchain
langchain-community
langchain-text-splitters
langchain-huggingface
huggingface-hub<=0.23.5
faiss-cpu
sentence-transformers
python-dotenv
streamlit
